+++ src/Cond.cc
@@ -39,7 +39,7 @@ void *Cond(void *argument)
   int i;
   double missval1 = -9.E33;
   double missval2 = -9.E33;
-  int **varnmiss1 = NULL;
+  size_t **varnmiss1 = NULL;
   double **vardata1 = NULL;
 
   cdoInitialize(argument);
@@ -103,13 +103,13 @@ void *Cond(void *argument)
 
 	  nvars  = vlistNvars(vlistID1);
 	  vardata1  = (double **) Malloc(nvars*sizeof(double *));
-	  varnmiss1 = (int **) Malloc(nvars*sizeof(int *));
+	  varnmiss1 = (size_t **) Malloc(nvars*sizeof(size_t *));
 	  for ( varID = 0; varID < nvars; varID++ )
 	    {
 	      gridsize = gridInqSize(vlistInqVarGrid(vlistID1, varID));
 	      nlev     = zaxisInqSize(vlistInqVarZaxis(vlistID1, varID));
 	      vardata1[varID]  = (double*) Malloc(nlev*gridsize*sizeof(double));
-	      varnmiss1[varID] = (int*) Malloc(nlev*sizeof(int));
+	      varnmiss1[varID] = (size_t*) Malloc(nlev*sizeof(size_t));
 	    }
 	}
     }
