+++ src/Intntime.cc
@@ -60,8 +60,8 @@ void *Intntime(void *argument)
   int gridsize = vlistGridsizeMax(vlistID1);
   double *array = (double*) Malloc(gridsize*sizeof(double));
 
-  int **nmiss1   = (int **) Malloc(nvars*sizeof(int *));
-  int **nmiss2   = (int **) Malloc(nvars*sizeof(int *));
+  size_t **nmiss1   = (size_t **) Malloc(nvars*sizeof(size_t *));
+  size_t **nmiss2   = (size_t **) Malloc(nvars*sizeof(size_t *));
   double **vardata1 = (double **) Malloc(nvars*sizeof(double *));
   double **vardata2 = (double **) Malloc(nvars*sizeof(double *));
 
@@ -69,8 +69,8 @@ void *Intntime(void *argument)
     {
       gridsize = gridInqSize(vlistInqVarGrid(vlistID1, varID));
       nlevel   = zaxisInqSize(vlistInqVarZaxis(vlistID1, varID));
-      nmiss1[varID]   = (int*) Malloc(nlevel*sizeof(int));
-      nmiss2[varID]   = (int*) Malloc(nlevel*sizeof(int));
+      nmiss1[varID]   = (size_t*) Malloc(nlevel*sizeof(size_t));
+      nmiss2[varID]   = (size_t*) Malloc(nlevel*sizeof(size_t));
       vardata1[varID] = (double*) Malloc(gridsize*nlevel*sizeof(double));
       vardata2[varID] = (double*) Malloc(gridsize*nlevel*sizeof(double));
     }
