+++ src/Vertintml.cc
@@ -186,13 +186,13 @@ void *Vertintml(void *argument)
 
   std::vector<bool> vars(nvars);
   std::vector<bool> varinterp(nvars);
-  std::vector<int *> varnmiss(nvars);
+  std::vector<size_t *> varnmiss(nvars);
   std::vector<double *> vardata1(nvars);
   std::vector<double *> vardata2(nvars);
 
   int maxlev = nhlevh > nplev ? nhlevh : nplev;
 
-  int *pnmiss = extrapolate ? NULL : (int*) Malloc(nplev*sizeof(int));
+  size_t *pnmiss = extrapolate ? NULL : (size_t*) Malloc(nplev*sizeof(size_t));
 
   // check levels
   if ( zaxisIDh != -1 )
@@ -362,8 +362,8 @@ void *Vertintml(void *argument)
 	{
 	  varinterp[varID] = true;
 	  vardata2[varID]  = (double*) Malloc(gridsize*nplev*sizeof(double));
-	  varnmiss[varID]  = (int*) Malloc(maxlev*sizeof(int));
-	  memset(varnmiss[varID], 0, maxlev*sizeof(int));
+	  varnmiss[varID]  = (size_t*) Malloc(maxlev*sizeof(size_t));
+	  memset(varnmiss[varID], 0, maxlev*sizeof(size_t));
 	}
       else
 	{
@@ -376,7 +376,7 @@ void *Vertintml(void *argument)
 
 	  varinterp[varID] = false;
 	  vardata2[varID]  = vardata1[varID];
-	  varnmiss[varID]  = (int*) Malloc(nlevel*sizeof(int));
+	  varnmiss[varID]  = (size_t*) Malloc(nlevel*sizeof(size_t));
 	}
     }
 
@@ -597,7 +597,7 @@ void *Vertintml(void *argument)
 			       vert_index, plev, nplev, gridsize, nlevel, missval);
 		    }
 		  
-		  if ( !extrapolate ) memcpy(varnmiss[varID], pnmiss, nplev*sizeof(int));
+		  if ( !extrapolate ) memcpy(varnmiss[varID], pnmiss, nplev*sizeof(size_t));
 		}
 	    }
 	}
